requirements.txt
streamlit
bs4
google-generativeai